scripts/requirements.txt
# WhisperX and dependencies
whisperx>=3.1.1
torch>=2.0.0
torchaudio>=2.0.0
faster-whisper>=0.10.0
pyannote.audio>=3.1.0
pydub>=0.25.1

# For development
python-dotenv>=1.0.0